divide Filter into filter.ts
